Changed structure of Messaging (quite a lot)
- Added some dto classes for conversion between JSON and POJO
- DAOs used in MessagingService.java are now closed right after being
used. They're not being stored in private properties anymore.
- Reduces some methods in MessagingService.java by combining and reusing
methods from other classes like MessageDAO.getById
- (important) Splited handleSendMessage and handleUnsendMessage of
MessageEndpoint.java to both it and MessagingService.java. The purpose
of this change is to make sure MessageEndpoint and MessagingService does
its intended purpose and nothing more.
  + Purpose of MessagingEndpoint: Handle WebSocket connections and
communications, which includes using Payloads and JSONs to communicate.
  + Purpose of MessagingService: Do the actual JSON parsing, reading
Payload data, then use that data to process and return JSON objects or
DTOs back to caller.
